chore(general ledger): rrv schedule related issues fixed [GCP-2941] (#5696)

+++ app/Models/RecurringVoucherSetupSchedule.php
@@ -209,7 +209,7 @@ class RecurringVoucherSetupSchedule extends Model
         'processDate' => 'datetime',
         'RRVcode' => 'string',
         'currencyID' => 'integer',
-        'amount' => 'integer',
+        'amount' => 'float',
         'rrvGeneratedYN' => 'integer',
         'stopYN' => 'integer',
         'documentSystemID' => 'integer',
